pyproject.toml
[project]
name = "insurance"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.8",
    "logfire[fastapi]>=3.25.0",
    "pandas>=2.3.1",
    "pyarrow>=21.0.0",
    "python-dotenv>=1.1.1",
    "scikit-learn>=1.7.1",
    "uvicorn[standard]>=0.35.0",
    "zenml[server]>=0.84.0",
]
